requirements.txt
environs==9.5.0
python-telegram-bot==13.12
requests==2.28.0
Pillow==9.3.0
